Add prompting mechanism for config values
This allows not keeping passwords and such in the config file.

E.g.: In `/etc/aptly-cli.conf`:

```yaml
---
:proto: http
:server: 10.3.0.46
:port: 8082
:debug: false
:username: ${PROMPT}
:password: ${PROMPT_PASSWORD}
```

The tool will prompt for the specified values -- e.g.:

```
$ aptly-cli version
Enter a value for username:
zane
Enter a value for password:
***************
{"Version"=>"0.9.7"}
```

This also eliminates a bunch of duplicated code in all of the command
classes by having them inherit from a common base class. This was
actually essential because otherwise, every required file would trigger
a config load and thus prompting for any needed values over and over
again.