README.md: Tweak Docker instructions (#99)
The most important change is changing the image name in the sample command from `aptly-cli` to `sepulworld/aptly-cli`, so that it uses the new image that you uploaded to Docker Hub. It won't work without the `sepulworld/` prefix unless the user has built the image locally and named it `aptly-cli`.

I also changed the command to use `/etc/aptly-cli.conf` as the config file and that way you don't need to have a blurb about `~/.config/aptly-cli/aptly-cli.conf` and the Docker instructions more closely resemble the `gem install` instructions.